requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart
moderngl
openai
weave
python-dotenv
lpips
torch
torchvision
